Make state server responsible for resetting test database
