+++ Dockerfile.State
@@ -7,14 +7,12 @@ FROM mcr.microsoft.com/dotnet/core/sdk:3.1 AS installer-env
 COPY pact/StateServer/StateServer.csproj ./pact/StateServer/
 COPY src/Models/Models.csproj ./src/Models/
 COPY src/Database/Database.csproj ./src/Database/
-COPY tests/API/Integration/Integration.csproj ./tests/API/Integration/
 
 RUN dotnet restore ./pact/StateServer/StateServer.csproj
 
 # then copy the rest and build/publish.
 COPY src ./src/
 COPY pact ./pact/
-COPY tests ./tests/
 
 RUN mkdir -p /home/site/wwwroot && \
     dotnet publish -c Release ./pact/StateServer/StateServer.csproj --output /home/site/wwwroot
@@ -23,6 +21,7 @@ FROM mcr.microsoft.com/azure-functions/dotnet:3.0-appservice
 ENV AzureWebJobsScriptRoot=/home/site/wwwroot
 ENV AzureFunctionsJobHost__Logging__Console__IsEnabled=true
 ENV AzureWebJobsDisableHomepage=true
+ENV DatabaseConnectionString=Server=integration-test-db;Database=ItPeople;User\ Id=SA;Password=abcd1234@;
 COPY --from=installer-env ["/home/site/wwwroot", "/home/site/wwwroot"]
 
 HEALTHCHECK --interval=10s --timeout=1s --retries=60 \
