+++ pact/StateServer/StateServer.csproj
@@ -5,9 +5,11 @@
   </PropertyGroup>
   <ItemGroup>
     <PackageReference Include="Microsoft.NET.Sdk.Functions" Version="3.0.11" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore" Version="3.1.11" />
   </ItemGroup>
   <ItemGroup>
-    <ProjectReference Include="../../tests/API/Integration/Integration.csproj" />
+    <ProjectReference Include="../../src/Models/Models.csproj" />
+    <ProjectReference Include="../../src/Database/Database.csproj" />
   </ItemGroup>
   <ItemGroup>
     <None Update="host.json">
