+++ tests/API/Integration/Scaffolding/DatabaseContainer.cs
@@ -38,106 +38,5 @@ namespace Integration
         }
 
         protected abstract DbConnection GetConnection();
-
-        public static void ResetDatabase()
-        {
-            using (var peopleContext = PeopleContext.Create(PeopleContext.LocalDatabaseConnectionString))
-            {
-                ResetTestData(peopleContext);
-            }
-        }
-
-        private static void ResetTestData(PeopleContext peopleContext)
-        {
-            peopleContext.Database.ExecuteSqlRaw(@"
-                TRUNCATE 
-                    public.buildings, 
-                    public.building_relationships, 
-                    public.departments,
-                    public.people, 
-                    public.support_relationships, 
-                    public.tools,
-                    public.units,
-                    public.unit_members,
-                    public.unit_member_tools
-                RESTART IDENTITY
-                CASCADE;
-            ");
-
-            peopleContext.Buildings.AddRange(new List<Building> {
-                TestEntities.Buildings.CityHall,
-                TestEntities.Buildings.RonsCabin,
-                TestEntities.Buildings.SmallPark,
-            });
-
-            peopleContext.BuildingRelationships.AddRange(new List<BuildingRelationship> {
-                TestEntities.BuildingRelationships.CityHallCityOfPawnee,
-                TestEntities.BuildingRelationships.RonsCabinCityOfPawnee,
-            });
-
-            peopleContext.Departments.AddRange(new List<Department> {
-                TestEntities.Departments.Parks,
-                TestEntities.Departments.Fire,
-                TestEntities.Departments.Auditor
-            });
-            
-            peopleContext.Units.AddRange(new List<Unit> {
-                TestEntities.Units.CityOfPawnee,
-                TestEntities.Units.ParksAndRecUnit,
-                TestEntities.Units.Auditor,
-            });
-
-            peopleContext.People.AddRange(new List<Person> { 
-                TestEntities.People.RSwanson, 
-                TestEntities.People.LKnope,
-                TestEntities.People.BWyatt,
-                TestEntities.People.ServiceAdmin
-            });  
-            
-            peopleContext.SupportRelationships.AddRange(new List<SupportRelationship> {
-                TestEntities.SupportRelationships.ParksAndRecRelationship,
-                TestEntities.SupportRelationships.ParksAndRecUnitFire
-
-            });
-
-            peopleContext.UnitMembers.AddRange(new List<UnitMember> { 
-                TestEntities.UnitMembers.RSwansonDirector,
-                TestEntities.UnitMembers.LkNopeSublead,
-                TestEntities.UnitMembers.BWyattAditor
-            });
-            
-            peopleContext.Tools.AddRange(new List<Tool> { 
-                TestEntities.Tools.Hammer
-            });
-
-            peopleContext.MemberTools.AddRange(new List<MemberTool> { 
-                TestEntities.MemberTools.MemberTool
-            });
-            
-            peopleContext.SaveChanges();
-
-            // Account for the identities of Units we've already added to the database.  Prevents error:
-            //      duplicate key value violates unique constraint
-            peopleContext.Database.ExecuteSqlRaw(@"
-                SELECT
-                    setval(pg_get_serial_sequence('public.units', 'id'), 
-                    (SELECT MAX(id) FROM public.units));
-                ");
-            peopleContext.Database.ExecuteSqlRaw(@"
-                SELECT
-                    setval(pg_get_serial_sequence('public.building_relationships', 'id'), 
-                    (SELECT MAX(id) FROM public.building_relationships));
-                ");
-            peopleContext.Database.ExecuteSqlRaw(@"
-                SELECT
-                    setval(pg_get_serial_sequence('public.support_relationships', 'id'), 
-                    (SELECT MAX(id) FROM public.support_relationships));
-                ");
-            peopleContext.Database.ExecuteSqlRaw(@"
-                SELECT
-                    setval(pg_get_serial_sequence('public.unit_members', 'id'), 
-                    (SELECT MAX(id) FROM public.unit_members));
-                ");
-        }
     }
 }
